+++ packages/intellisense/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./conflicts";
 export * from "./constants";
+export type { IntellisenseConfig } from "./core";
 export * from "./core";
 export * from "./hover";
 export * from "./sort";